BF: Fix style of frontend build and allow the frontend to run alone
The style of the build-page differed from the style of the development-page. With the current fix the build-page now looks like the development-page again.

The commit also removes superfluous code in the css file and contains a fix to avoid a crash when the frontend is run without an active backend running.